+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -182,7 +182,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-28T15:44:09.058Z",
+    "2025-10-29T04:55:50.891Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -223,7 +223,7 @@ export const CreateContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-28T15:44:09.058Z"))
+  updated: z.date().default(() => new Date("2025-10-29T04:55:50.891Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {