+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -301,7 +301,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-28T15:44:09.058Z",
+    "2025-10-29T04:55:50.891Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -341,7 +341,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-28T15:44:09.058Z"))
+  updated: z.date().default(() => new Date("2025-10-29T04:55:50.891Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {